pyproject.toml
[project]
name = "pollen-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jay Gaha", email = "gaha@buzzreach.co.jp" }
]
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
